Utilize useMemo to rerender BAC and remove drinks state type redundancy
Drinks State was of type <Spec[]> similar to Cocktail state. Although benefitial previously, we updated our server to handle a lot of these actions.

New Approach consits of having a Drinks state with type <DrinksCocktail[]> and Cocktail state with type <Specs[]> and remove totalEthanol state and reduce that state from Drinks

Step 1: Integrating it in BAC and AddEthanolForm

Next steps include implementing it in the custom hook which handles Dilution and nUtrition, Accepting Drinks data to DB instead of ethanol and finally utilizing this instead of ethanol